// types/_dashboard/rentcast_types.ts
type PropertyType =
	| "Single Family"
	| "Multi-Family"
	| "Condo"
	| "Townhouse"
	| "Apartment"
	| "Mobile Home"
	| "Land"
	| "Other";